requirements/ginkgo.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=ginkgo.txt ginkgo.in
#
astroid==1.5.3            # via pylint
backports.functools-lru-cache==1.5  # via astroid, isort, pylint
bumpversion==0.5.3
cachetools==3.1.1         # via google-auth
configparser==3.7.4       # via pylint
django==1.11.22
djangorestframework==3.6.3
edx-opaque-keys==0.4.4
enum34==1.1.6             # via astroid
futures==3.3.0            # via isort
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3
httplib2==0.13.0          # via google-api-python-client, google-auth-httplib2
isort==4.3.21             # via pylint
lazy-object-proxy==1.4.1  # via astroid
mccabe==0.6.1             # via pylint
pbr==5.4.1                # via stevedore
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pycodestyle==2.5.0
pylint==1.7.6
pymongo==3.8.0            # via edx-opaque-keys
pytz==2019.1              # via django
rsa==4.0                  # via google-auth
singledispatch==3.4.0.3   # via astroid, pylint
six==1.12.0               # via astroid, edx-opaque-keys, google-api-python-client, google-auth, pylint, singledispatch, stevedore
stevedore==1.30.1         # via edx-opaque-keys
uritemplate==3.0.0        # via google-api-python-client
wrapt==1.11.2             # via astroid
